# Remove seemingly unneeded deepcopy's in macro expansion.