Timings using test_loader.py.

Before:

```
load_with_references (/Users/john/workspace/galaxy/lib/galaxy/util/xml_macros.py:11)
function called 14 times

         26836 function calls (23318 primitive calls) in 0.028 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 121 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       14    0.000    0.000    0.028    0.002 xml_macros.py:11(load_with_references)
    44/14    0.000    0.000    0.018    0.001 xml_macros.py:127(_expand_macros)
    30/27    0.000    0.000    0.017    0.001 xml_macros.py:141(_expand_macro)
  1608/78    0.004    0.000    0.010    0.000 copy.py:145(deepcopy)
```

After:

```
load_with_references (/Users/john/workspace/galaxy/lib/galaxy/util/xml_macros.py:11)
function called 14 times

         18269 function calls (15767 primitive calls) in 0.020 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 121 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       14    0.000    0.000    0.020    0.001 xml_macros.py:11(load_with_references)
    44/14    0.000    0.000    0.011    0.001 xml_macros.py:127(_expand_macros)
    30/27    0.000    0.000    0.009    0.000 xml_macros.py:141(_expand_macro)
       15    0.000    0.000    0.007    0.000 xml_macros.py:311(_parse_xml)
       14    0.000    0.000    0.007    0.000 xml_macros.py:51(raw_xml_tree)
       15    0.000    0.000    0.006    0.000 ElementTree.py:1180(parse)
       15    0.000    0.000    0.006    0.000 ElementTree.py:644(parse)
   672/30    0.002    0.000    0.005    0.000 copy.py:145(deepcopy)
```

Also add more tests to try to verify the deepcopys are not needed.