@wm75 2019-08-07
@jmchilton if you're looking for more test cases, you may want to run the tests for bcftools_mpileup against your branch. There is a nested token expansion in it that resulted in a really hard to reproduce error (dependent on the actual *names* of the tokens) for me just recently (see https://travis-ci.org/galaxyproject/tools-iuc/jobs/541356154 and https://github.com/galaxyproject/tools-iuc/pull/2440#issuecomment-498829504 and below). Maybe you want to check that your change doesn't make this problem reappear?