@mvdbeek 2019-08-09
@wm75 I ran the bcftools mpileup tests against this branch, it's all working fine.